First commit for #18. Framework for game logic.
